requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.6
django-cors-headers==4.6.0
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.10
importlib==1.0.4
packaging==24.2
pillow==11.1.0
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
whitenoise==6.9.0
